Allow RPMs to be downgraded within Ansible
Ansible is not currently able to install an older version of an RPM (if
this older version was previously installed on the system) because
of the way yum handles downgrading RPMs.

In order to downgrade to an older version of an RPM, one needs to use
the `yum downgrade` plugin, since the intuitive `yum install <older
version>` seems too simple in a CentOS world.

As of now, yum downgrading _will not_ work on lists of packages. This is
a necessary workaround for now since building the logic for this will
take a larger refactoring effort.

This work builds on the work done in previous PRs:

- https://github.com/ansible/ansible/pull/6033
- https://github.com/ansible/ansible/pull/6333